requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
openai==1.30.0
python-dotenv==1.0.1
numpy==1.26.4
pytest==8.2.0
httpx==0.27.0
